feat(plasma-new-hope): Add and refactoring icons
