+++ packages/plasma-new-hope/src/components/_Icon/Icon.assets/Done.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { IconProps } from '../IconRoot';
+
+export const Done: React.FC<IconProps> = (props) => (
+    <svg width="100%" viewBox="0 0 24 24" fill="none" {...props}>
+        <path
+            d="M9 16.2l-3.5-3.5a.984.984 0 00-1.4 0 .984.984 0 000 1.4l4.19 4.19c.39.39 1.02.39 1.41 0L20.3 7.7a.984.984 0 000-1.4.984.984 0 00-1.4 0L9 16.2z"
+            fill="currentColor"
+        />
+    </svg>
+);
