+++ packages/plasma-new-hope/src/components/_Icon/Icons/IconChevronLeft.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { ChevronLeft } from './Icon.assets/ChevronLeft';
-import { IconRoot, IconProps } from './IconRoot';
+import { ChevronLeft } from '../Icon.assets/ChevronLeft';
+import { IconRoot, IconProps } from '../IconRoot';
 
 export const IconChevronLeft: React.FC<IconProps> = ({ size = 's', color, className, sizeCustomProperty }) => {
     return (
