+++ packages/plasma-new-hope/src/components/_Icon/Icons/IconClose.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Close } from './Icon.assets/Close';
-import { IconRoot, IconProps } from './IconRoot';
+import { Close } from '../Icon.assets/Close';
+import { IconRoot, IconProps } from '../IconRoot';
 
 export const IconClose: React.FC<IconProps> = ({ size = 's', color, className, sizeCustomProperty }) => {
     return (
