+++ packages/plasma-new-hope/src/components/_Icon/Icons/IconCross.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Cross } from './Icon.assets/Cross';
-import { IconRoot, IconProps } from './IconRoot';
+import { Cross } from '../Icon.assets/Cross';
+import { IconRoot, IconProps } from '../IconRoot';
 
 export const IconCross: React.FC<IconProps> = ({ size = 's', color, className, sizeCustomProperty }) => {
     return (
