+++ packages/plasma-new-hope/src/components/_Icon/Icons/IconDisclosureLeft.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { DisclosureLeft } from './Icon.assets/DisclosureLeft';
-import { IconRoot, IconProps } from './IconRoot';
+import { DisclosureLeft } from '../Icon.assets/DisclosureLeft';
+import { IconRoot, IconProps } from '../IconRoot';
 
 export const IconDisclosureLeft: React.FC<IconProps> = ({ size = 's', color, className }) => {
     return <IconRoot className={className} size={size} color={color} icon={DisclosureLeft} />;
