+++ packages/plasma-new-hope/src/components/_Icon/Icons/IconDone.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { Done } from '../Icon.assets/Done';
+import { IconRoot, IconProps } from '../IconRoot';
+
+export const IconDone: React.FC<IconProps> = ({ size = 's', color, className }) => {
+    return <IconRoot className={className} size={size} color={color} icon={Done} />;
+};
