+++ packages/plasma-new-hope/src/components/_Icon/Icons/IconMic.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Mic } from './Icon.assets/Mic';
-import { IconRoot, IconProps } from './IconRoot';
+import { Mic } from '../Icon.assets/Mic';
+import { IconRoot, IconProps } from '../IconRoot';
 
 export const IconMic: React.FC<IconProps> = ({ size = 's', color, className }) => {
     return <IconRoot className={className} size={size} color={color} icon={Mic} />;
