+++ packages/plasma-new-hope/src/components/_Icon/index.tsx
@@ -1,4 +1,8 @@
-export { IconDisclosureLeft } from './IconDisclosureLeft';
-export { IconDisclosureRight } from './IconDisclosureRight';
-export { IconMic } from './IconMic';
-export { IconChevronLeft } from './IconChevronLeft';
+export { IconDisclosureLeft } from './Icons/IconDisclosureLeft';
+export { IconDisclosureRight } from './Icons/IconDisclosureRight';
+export { IconMic } from './Icons/IconMic';
+export { IconChevronLeft } from './Icons/IconChevronLeft';
+export { IconChevronDown } from './Icons/IconChevronDown';
+export { IconClose } from './Icons/IconClose';
+export { IconCross } from './Icons/IconCross';
+export { IconDone } from './Icons/IconDone';
